+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * `filter(HashLike, predicate)` - the `predicate` doesn't have to be `Fun` anymore
 * `filter(hl:HashLike, predicate)` now returns value of the same type as `hl`
 * `escape_bash()` - fixed and it's not experimental anymore
+* `section` inside `ns` now working properly
 
 ### Breaking changes